Commenting, anglicizing and debugging
Finished commenting on forms.js. Continued commenting on add_form.php and edit_form.php

Continued anglicizing some local variables (but it proved a bit tricky for frequently used variables such as "etat", so I left it as is)

Debugged the issue that photos of items updated in edit_form.php where not immediately shown because the cache was not updated. Now, I force update of the item photo in cache in item_page whenever the photo has just been updated, and the cache is always refreshed for the photo in edit_form.php.
Ideally, this should be resolved by having unique filenames for each new uploaded photos, and storing the photo URL in the database, rather than relying on the ID of the item to locate the photo URL.